+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 
 ### New features
 
+- Added `\Wizaplace\SDK\Basket\BasketService::setUserBasketId`
+- Added `\Wizaplace\SDK\Basket\BasketService::getUserBasketId`
 - Added `\Wizaplace\SDK\Catalog\Product::getAttachments`
 
 ### Bugfixes